@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 
 class UnknownScreen extends StatelessWidget {
-  const UnknownScreen({Key? key}) : super(key: key);
+  const UnknownScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(title: const Text('Opps')),
-      body: Column(
-        children: const [Text('Opps')],
+      body: const Column(
+        children: [Text('Opps')],
       ),
     );
   }
