@@ -1,9 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:untitled/base/dependency/i18n/i18n_base_provider.dart';
 
-class I18nProvider extends I18nBaseProvider {
-  @override
+class I18nProvider {
   AppLocalizations localizationsInstance(BuildContext context) {
     return AppLocalizations.of(context)!;
   }
