@@ -1,10 +1,10 @@
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:untitled/base/dependency/i18n/i18n_base_provider.dart';
+import 'package:untitled/base/dependency/i18n/i18n_provider.dart';
 
-import '../router/router_provider_interface.dart';
+import 'package:untitled/base/dependency/router/router_provider_interface.dart';
 
 class I18nService {
-  final I18nBaseProvider i18nProvider;
+  final I18nProvider i18nProvider;
   final RouterProviderInterface routerProvider;
 
   I18nService({
