@@ -1,7 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../base/bloc/bloc_base.dart';
-import '../../packages/rx/main.dart';
+import 'package:untitled/base/bloc/bloc_base.dart';
+import 'package:untitled/packages/rx/main.dart';
 
 class CounterBloc extends BlocBase {
   final counterSubject = BehaviorSubject.seeded(0);
